chore: fix stuff

+++ barretenberg/cpp/src/barretenberg/dsl/acir_format/ecdsa_secp256r1.test.cpp
@@ -27,29 +27,29 @@ size_t generate_r1_constraints(EcdsaSecp256r1Constraint& ecdsa_r1_constraint,
     std::array<uint32_t, 32> pub_y_indices_in;
     std::array<uint32_t, 64> signature_in;
     size_t offset = 0;
-    for (size_t i = 0; i < hashed_message.size(); ++i) {
+    for (uint32_t i = 0; i < hashed_message.size(); ++i) {
         message_in[i] = (i + offset);
         const auto byte = static_cast<uint8_t>(hashed_message[i]);
         witness_values.emplace_back(byte);
     }
     offset += message_in.size();
 
-    for (size_t i = 0; i < 32; ++i) {
+    for (uint32_t i = 0; i < 32; ++i) {
         pub_x_indices_in[i] = (i + offset);
         witness_values.emplace_back(pub_x_value.slice(248 - i * 8, 256 - i * 8));
     }
     offset += pub_x_indices_in.size();
-    for (size_t i = 0; i < 32; ++i) {
+    for (uint32_t i = 0; i < 32; ++i) {
         pub_y_indices_in[i] = (i + offset);
         witness_values.emplace_back(pub_y_value.slice(248 - i * 8, 256 - i * 8));
     }
     offset += pub_y_indices_in.size();
-    for (size_t i = 0; i < 32; ++i) {
+    for (uint32_t i = 0; i < 32; ++i) {
         signature_in[i] = (i + offset);
         witness_values.emplace_back(signature.r[i]);
     }
     offset += signature.r.size();
-    for (size_t i = 0; i < 32; ++i) {
+    for (uint32_t i = 0; i < 32; ++i) {
         signature_in[i] = (i + offset);
         witness_values.emplace_back(signature.s[i]);
     }
